rewire things and fix issues
