+++ cli/main.rs
@@ -121,6 +121,7 @@ fn create_web_worker_callback(
       no_color: !colors::use_color(),
       get_error_class_fn: Some(&crate::errors::get_error_class_name),
       blob_url_store: program_state.blob_url_store.clone(),
+      startup_snapshot: Some(deno_snapshots::deno_isolate_init()),
     };
 
     let mut worker = WebWorker::from_options(
@@ -201,6 +202,7 @@ pub fn create_main_worker(
     get_error_class_fn: Some(&crate::errors::get_error_class_name),
     location: program_state.flags.location.clone(),
     blob_url_store: program_state.blob_url_store.clone(),
+    startup_snapshot: Some(deno_snapshots::deno_isolate_init()),
   };
 
   let mut worker = MainWorker::from_options(main_module, permissions, &options);
