+++ cli/tsc.rs
@@ -20,7 +20,6 @@ use deno_core::JsRuntime;
 use deno_core::ModuleSpecifier;
 use deno_core::OpFn;
 use deno_core::RuntimeOptions;
-use deno_core::Snapshot;
 use std::collections::HashMap;
 use std::path::PathBuf;
 use std::sync::Arc;
@@ -42,13 +41,6 @@ pub static SHARED_GLOBALS_LIB: &str =
   include_str!("dts/lib.deno.shared_globals.d.ts");
 pub static WINDOW_LIB: &str = include_str!("dts/lib.deno.window.d.ts");
 pub static UNSTABLE_NS_LIB: &str = include_str!("dts/lib.deno.unstable.d.ts");
-
-pub static COMPILER_SNAPSHOT: &[u8] =
-  include_bytes!(concat!(env!("OUT_DIR"), "/COMPILER_SNAPSHOT.bin"));
-
-pub fn compiler_snapshot() -> Snapshot {
-  Snapshot::Static(COMPILER_SNAPSHOT)
-}
 
 macro_rules! inc {
   ($e:expr) => {
@@ -444,7 +436,7 @@ fn op_respond(state: &mut State, args: Value) -> Result<Value, AnyError> {
 /// optionally an updated TypeScript build info.
 pub fn exec(request: Request) -> Result<Response, AnyError> {
   let mut runtime = JsRuntime::new(RuntimeOptions {
-    startup_snapshot: Some(compiler_snapshot()),
+    startup_snapshot: Some(deno_snapshots::compiler_snapshot()),
     ..Default::default()
   });
   // tsc cannot handle root specifiers that don't have one of the "acceptable"
@@ -618,7 +610,7 @@ mod tests {
   #[test]
   fn test_compiler_snapshot() {
     let mut js_runtime = deno_core::JsRuntime::new(deno_core::RuntimeOptions {
-      startup_snapshot: Some(compiler_snapshot()),
+      startup_snapshot: Some(deno_snapshots::compiler_snapshot()),
       ..Default::default()
     });
     js_runtime
